requirements.txt
python-chess >= 0.26.0
pickledb >= 0.9.2
tensorflow >= 1.13.1
keras >= 2.2.4
numpy >= 1.16.2
matplotliB >= 3.0.3
stockfish >= 1.0.4
h5py >= 2.9.0
#python3-tk
#python-pydot
#python-pydot-ng
graphviz >= 0.10.1

